requirements.txt
keyboard>=0.13.5
psutil>=5.9.0
pystray>=0.19.0
Pillow>=9.0.0
sounddevice>=0.4.5
numpy>=1.21.0
pydub>=0.25.1
pyaudiowpatch>=0.2.1
